Start on splitting Java and Scala ResourceManglers
