+++ src/main/java/com/meteorcode/pathway/io/java_api/FilesystemFileHandle.scala
@@ -1,6 +1,7 @@
 package com.meteorcode.pathway.io.java_api
 
-import com.meteorcode.pathway.io.{ResourceManager, scala_api}
+import com.meteorcode.pathway.io.scala_api
+import com.meteorcode.pathway.io.scala_api.ResourceManager
 
 /**
  * Wrapper to give Java callers a constructor for creating initial seed [[FileHandle]]s.
@@ -9,6 +10,7 @@ import com.meteorcode.pathway.io.{ResourceManager, scala_api}
 class FilesystemFileHandle private[this](underlying: scala_api.FileHandle) extends FileHandle(underlying) {
 
   def this(virtualPath: String, realPath: String, manager: ResourceManager) = this(
-    new scala_api.FilesystemFileHandle(virtualPath, realPath, manager)
+    new FilesystemFileHandle(virtualPath, realPath, manager)
   )
 }
+
