+++ src/main/java/com/meteorcode/pathway/io/scala_api/FileHandle.scala
@@ -3,7 +3,7 @@ package com.meteorcode.pathway.io.scala_api
 import java.io.{BufferedInputStream, BufferedOutputStream, File, IOException, InputStream, OutputStream}
 import java.nio.charset.Charset
 
-import com.meteorcode.pathway.io.ResourceManager
+import com.meteorcode.pathway.io.scala_api.ResourceManager
 
 import scala.io.Source
 import scala.util.{Try, Success, Failure}
