+++ src/test/java/com/meteorcode/pathway/test/IOSpec.scala
@@ -3,10 +3,8 @@ package com.meteorcode.pathway.test
 import java.io._
 import java.nio.charset.Charset
 import java.nio.file.{FileSystems, Files}
-
-import com.meteorcode.pathway.io.ResourceManager
 import com.meteorcode.pathway.io.java_api.AlphabeticLoadPolicy
-import com.meteorcode.pathway.io.scala_api.{FilesystemFileHandle, FileHandle}
+import com.meteorcode.pathway.io.scala_api.{ResourceManager, FilesystemFileHandle, FileHandle}
 import com.meteorcode.pathway.test.tags.FilesystemTest
